Web mp3 stream.... sorta
This is pretty wrong. The whole song is sent to the client. If the
server is playing, then... cool hopefully the mpg123 player terminates
at the same time. If so then it kinda works.

If the mpg123 process terms early e.g. skip. The web stream is totally
unaware and continues to stream. So there needs to be some coordination
between the NextSong loop and anyone who is streaming.